pyproject.toml
[project]
name = "country-currency-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
